step5: move FixedStrategy.kt to test package
