+++ src/test/kotlin/step4/FixedStrategy.kt
@@ -1,4 +1,6 @@
-package step4.domain
+package step4
+
+import step4.domain.MoveStrategy
 
 class FixedStrategy(override val count: Int) : MoveStrategy {
     override fun isMoveSuccess(): Boolean {
